Add support for 64-bit scripts;

+++ XtractQuery/Logic.Domain.Level5/Logic.Domain.Level5.Contract/Script/Xq32/DataClasses/Xq32Function.cs
@@ -2,7 +2,7 @@
 {
     public struct Xq32Function
     {
-        public int nameOffset;
+        public long nameOffset;
         public uint crc32;
         public short instructionOffset;
         public short instructionEndOffset;
