+++ XtractQuery/Logic.Domain.Level5/Logic.Domain.Level5.Contract/Script/Xq32/DataClasses/Xq32Jump.cs
@@ -2,7 +2,7 @@
 {
     public struct Xq32Jump
     {
-        public int nameOffset;
+        public long nameOffset;
         public uint crc32;
         public int instructionIndex;
     }
