+++ XtractQuery/Logic.Domain.Level5/Logic.Domain.Level5.Contract/Script/Xseq/DataClasses/XseqJump.cs
@@ -2,7 +2,7 @@
 {
     public struct XseqJump
     {
-        public int nameOffset;
+        public long nameOffset;
         public ushort crc16;
         public short instructionIndex;
     }
